requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
# brotlipy==0.7.0
cachetools==5.3.1
click==8.1.7
datasets==2.14.5
decorator==4.4.2
dill==0.3.7
exceptiongroup==1.1.3
fastapi==0.103.1
Flask==2.3.3
frozenlist==1.4.0
fsspec==2023.6.0
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-speech==2.21.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.17.1
imageio==2.31.3
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
itsdangerous==2.1.2
joblib==1.3.2
moviepy==1.0.3
multidict==6.0.4
multiprocess==0.70.15
nltk==3.8.1
oauthlib==3.2.2
openai==0.28.0
packaging==23.1
pandas==2.1.0
Pillow==9.4.0
proglog==0.1.10
proto-plus==1.22.3
protobuf==4.24.3
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.2
tzdata==2023.3
uritemplate==4.1.1
uvicorn==0.23.2
Werkzeug==2.3.7
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
